fix(src): Sistemati Metodi Dim Reduction
Aggiornati metodi dim reduction per rispecchiare la nuova
struttura della libreria.

Aggiornata documentazione.
