+++ docs/index.html
@@ -37,7 +37,7 @@ import cnlp.utils</code></pre>
 <section>
 <h2 class="section-title" id="header-submodules">Sub-modules</h2>
 <dl>
-<dt><code class="name"><a title="cnlp.dimensionality_reduction_methods" href="dimensionality_reduction_methods/index.html">cnlp.dimensionality_reduction_methods</a></code></dt>
+<dt><code class="name"><a title="cnlp.dimensionality_reduction_methods" href="dimensionality_reduction_methods.html">cnlp.dimensionality_reduction_methods</a></code></dt>
 <dd>
 <div class="desc"><p>Dimentionality Reduction based Methods for Link Prediction …</p></div>
 </dd>
@@ -74,7 +74,7 @@ import cnlp.utils</code></pre>
 <ul id="index">
 <li><h3><a href="#header-submodules">Sub-modules</a></h3>
 <ul>
-<li><code><a title="cnlp.dimensionality_reduction_methods" href="dimensionality_reduction_methods/index.html">cnlp.dimensionality_reduction_methods</a></code></li>
+<li><code><a title="cnlp.dimensionality_reduction_methods" href="dimensionality_reduction_methods.html">cnlp.dimensionality_reduction_methods</a></code></li>
 <li><code><a title="cnlp.other_methods" href="other_methods/index.html">cnlp.other_methods</a></code></li>
 <li><code><a title="cnlp.probabilistic_methods" href="probabilistic_methods/index.html">cnlp.probabilistic_methods</a></code></li>
 <li><code><a title="cnlp.similarity_methods" href="similarity_methods/index.html">cnlp.similarity_methods</a></code></li>
